Usability adjustments
Call with input and output filenames.
For ancillary files, put modifiers in the basename of the file, not the extension.
Report the filenames for Log and Err that the script will create.
Use the new versions of oplsfm.pl and de_oplsfm.pl.
Delete the Err file if it is empty.